StudProjects/team04/project/requirements.txt
flask-cors==3.0.8
tensorflow==2.0.0
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
dependency-injector==3.14.12
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.5
MarkupSafe==1.1.1
monkey==0.1
more-itertools==7.2.0
mss==4.0.3
numpy==1.17.3
opencv-python==4.1.1.26
pycparser==2.19
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
